zoo changed from double to int

+++ src/main/java/dataaccessinterface/TileRepository.java
@@ -5,7 +5,7 @@ import entity.WeatherTile;
 import java.awt.image.BufferedImage;
 
 public interface TileRepository {
-    BufferedImage getTileImageData(int x, int y, double zoom, java.time.Instant timestamp) throws TileNotFoundException;
+    BufferedImage getTileImageData(int x, int y, int zoom, java.time.Instant timestamp) throws TileNotFoundException;
     BufferedImage getTileImageData(WeatherTile tile) throws TileNotFoundException;
     void addTileToCache(WeatherTile tile) throws TileNotFoundException;
 }
